Membuat Class baru
Membuat Class Child, SalaryEmployee
