+++ src/lib/Child.java
@@ -1,0 +1,16 @@
+package lib;
+
+public class Child extends Employee{
+   
+	private List<String> childNames;
+	private List<String> childIdNumbers;
+
+    public Child (){
+
+    }
+
+    public void addChild(String childName, String childIdNumber) {
+		childNames.add(childName);
+		childIdNumbers.add(childIdNumber);
+	}
+}
