+++ src/lib/Employee.java
@@ -21,15 +21,15 @@ public class Employee {
 	private boolean isForeigner;
 	private boolean gender; //true = Laki-laki, false = Perempuan
 	
-	private int monthlySalary;
-	private int otherMonthlyIncome;
-	private int annualDeductible;
+	// private int monthlySalary;
+	// private int otherMonthlyIncome;
+	// private int annualDeductible;
 	
 	private String spouseName;
 	private String spouseIdNumber;
 
-	private List<String> childNames;
-	private List<String> childIdNumbers;
+	// private List<String> childNames;
+	// private List<String> childIdNumbers;
 	
 	public Employee(String employeeId, String firstName, String lastName, String idNumber, String address, int yearJoined, int monthJoined, int dayJoined, boolean isForeigner, boolean gender) {
 		this.employeeId = employeeId;
@@ -43,51 +43,51 @@ public class Employee {
 		this.isForeigner = isForeigner;
 		this.gender = gender;
 		
-		childNames = new LinkedList<String>();
-		childIdNumbers = new LinkedList<String>();
+		// childNames = new LinkedList<String>();
+		// childIdNumbers = new LinkedList<String>();
 	}
 	
-	/**
-	 * Fungsi untuk menentukan gaji bulanan pegawai berdasarkan grade kepegawaiannya (grade 1: 3.000.000 per bulan, grade 2: 5.000.000 per bulan, grade 3: 7.000.000 per bulan)
-	 * Jika pegawai adalah warga negara asing gaji bulanan diperbesar sebanyak 50%
-	 */
+	// /**
+	//  * Fungsi untuk menentukan gaji bulanan pegawai berdasarkan grade kepegawaiannya (grade 1: 3.000.000 per bulan, grade 2: 5.000.000 per bulan, grade 3: 7.000.000 per bulan)
+	//  * Jika pegawai adalah warga negara asing gaji bulanan diperbesar sebanyak 50%
+	//  */
 	
-	public void setMonthlySalary(int grade) {	
-		if (grade == 1) {
-			monthlySalary = 3000000;
-			if (isForeigner) {
-				monthlySalary = (int) (3000000 * 1.5);
-			}
-		}else if (grade == 2) {
-			monthlySalary = 5000000;
-			if (isForeigner) {
-				monthlySalary = (int) (3000000 * 1.5);
-			}
-		}else if (grade == 3) {
-			monthlySalary = 7000000;
-			if (isForeigner) {
-				monthlySalary = (int) (3000000 * 1.5);
-			}
-		}
-	}
+	// public void setMonthlySalary(int grade) {	
+	// 	if (grade == 1) {
+	// 		monthlySalary = 3000000;
+	// 		if (isForeigner) {
+	// 			monthlySalary = (int) (3000000 * 1.5);
+	// 		}
+	// 	}else if (grade == 2) {
+	// 		monthlySalary = 5000000;
+	// 		if (isForeigner) {
+	// 			monthlySalary = (int) (3000000 * 1.5);
+	// 		}
+	// 	}else if (grade == 3) {
+	// 		monthlySalary = 7000000;
+	// 		if (isForeigner) {
+	// 			monthlySalary = (int) (3000000 * 1.5);
+	// 		}
+	// 	}
+	// }
 	
-	public void setAnnualDeductible(int deductible) {	
-		this.annualDeductible = deductible;
-	}
+	// public void setAnnualDeductible(int deductible) {	
+	// 	this.annualDeductible = deductible;
+	// }
 	
-	public void setAdditionalIncome(int income) {	
-		this.otherMonthlyIncome = income;
-	}
+	// public void setAdditionalIncome(int income) {	
+	// 	this.otherMonthlyIncome = income;
+	// }
 	
 	public void setSpouse(String spouseName, String spouseIdNumber) {
 		this.spouseName = spouseName;
 		this.spouseIdNumber = idNumber;
 	}
 	
-	public void addChild(String childName, String childIdNumber) {
-		childNames.add(childName);
-		childIdNumbers.add(childIdNumber);
-	}
+	// public void addChild(String childName, String childIdNumber) {
+	// 	childNames.add(childName);
+	// 	childIdNumbers.add(childIdNumber);
+	// }
 	
 	public int getAnnualIncomeTax() {
